move fullScreenButton to be shared, remove md settings
